+++ src/main/java/chronosacaria/mcdw/enums/IMcdwWeaponID.java
@@ -33,7 +33,7 @@ public interface IMcdwWeaponID {
 
     Item getItem();
 
-    Float getItemSpawnRate();
+    Integer getItemSpawnRate();
 
     Item makeWeapon();
 }
